[框架]添加断开连接 服务器读写

+++ Client/Assets/Scripts/Framework/Network/Define.cs
@@ -10,6 +10,6 @@ namespace Framework.Network {
     
     public static class ConnectConfig {
         public static int MaxReconnectCount = 5;
-        public static int ConnectTimeout = 10;
+        public static float ConnectTimeout = 10;
     }
 }
